Main Script
First, the data is imported and 90% of the training data is picked randomly in each run and used to train a random forest, the remaining 10% is used as validation set to calculate RMSE and MAE.

I decided to use Random Forest Estimator because the features of the data were diverse and discrete.
Using scikit-learn library of Python, the RandomForestRegressor() is trained on the training data. Grid Search of scikit-learn is used to optimize number of trees and maximum depth parameters. Number of learners is picked as 150 and maximum depth is 12. The rest is left as the default parameters. Grid search part is commented out in the code.